+++ Controls/_filterPanel/Editors/ListBase.wml
@@ -1,11 +1,10 @@
 <div class="controls-FilterEditors__list controls-FilterEditors__list_theme-{{_options.theme}}">
    <Controls.grid:View source="{{_options.source}}"
                        columns="{{_columns}}"
-                       markedKey="{{_options.propertyValue}}"
                        keyProperty="{{_options.keyProperty}}"
+                       markedKey="{{_options.propertyValue}}"
                        scope="{{_options}}"
                        navigation="{{_options.navigation}}"
-                       markerVisibility="{{false}}"
                        backgroundStyle="transparent"
                        hoverBackgroundStyle="listEditor"
                        on:markedKeyChanged="_handleMarkedKeyChanged()">
