+++ Controls/filterPanel.ts
@@ -9,4 +9,5 @@
 export {default as View} from './_filterPanel/View';
 export {default as NumberRangeEditor} from './_filterPanel/Editors/NumberRange';
 export {default as EnumListEditor} from './_filterPanel/Editors/EnumList';
+export {default as ListBase} from './_filterPanel/Editors/ListBase';
 export {default as Container} from './_filterPanel/View/Container';
